requirements.txt
Flask==2.2.5
Pillow==10.3.0
boto3==1.21.5